Phase 2: Document model, undo/redo, selection, snapping, and file I/O
Integrate DraftDocument/DraftEntity system with the UI, replacing raw
geometry storage in ViewportWidget with a proper document model.

- Add Document class (owns DraftDocument + UndoStack + dirty/path state)
- Add UndoStack with Command pattern (push/undo/redo)
- Add AddEntityCommand and RemoveEntityCommand
- Unify SelectionManager IDs to uint64_t to match DraftEntity
- Rewrite ViewportWidget to render from Document entities with selection
  highlighting (orange) and snap indicator (yellow cross)
- Wire LineTool and CircleTool to create entities via undo commands
- Implement SelectTool with zoom-aware hit tolerance, Shift+multi-select,
  and Delete key removal
- Implement NativeFormat JSON serialization (.hcad) for save/load
- Wire MainWindow: undo/redo, new/open/save/save-as, tool switching
- Fix double Y-flip bug in worldPositionAtCursor (screenToRay already
  handles Qt-to-NDC Y conversion)
- Add pixelToWorldScale() for screen-space-aware pick tolerance

+++ src/render/include/horizon/render/SelectionManager.h
@@ -12,19 +12,19 @@ public:
     SelectionManager();
     ~SelectionManager();
 
-    void select(uint32_t id);
-    void deselect(uint32_t id);
-    void toggle(uint32_t id);
+    void select(uint64_t id);
+    void deselect(uint64_t id);
+    void toggle(uint64_t id);
     void clearSelection();
 
-    bool isSelected(uint32_t id) const;
+    bool isSelected(uint64_t id) const;
 
-    std::vector<uint32_t> selectedIds() const;
+    std::vector<uint64_t> selectedIds() const;
     size_t count() const { return m_selected.size(); }
     bool empty() const { return m_selected.empty(); }
 
 private:
-    std::set<uint32_t> m_selected;
+    std::set<uint64_t> m_selected;
 };
 
 }  // namespace hz::render
